Module google_chromeuxreport1::api

Structs

Bin
A bin is a discrete portion of data spanning from start to end, or if no end is given, then from start to +inf. A bin's start and end values are given in the value type of the metric it represents. For example, "first contentful paint" is measured in milliseconds and exposed as ints, therefore its metric bins will use int32s for its start and end types. However, "cumulative layout shift" is measured in unitless decimals and is exposed as a decimal encoded as a string, therefore its metric bins will use strings for its value type.

ChromeUXReport
Central instance to access all ChromeUXReport related resource activities

Key
Key defines all the dimensions that identify this record as unique.

Metric
A metric is a set of user experience data for a single web performance metric, like "first contentful paint". It contains a summary histogram of real world Chrome usage as a series of bins.

Percentiles
Percentiles contains synthetic values of a metric at a given statistical percentile. These are used for estimating a metric's value as experienced by a percentage of users out of the total number of users.

QueryRequest
Request payload sent by a physical web client. This request includes all necessary context to load a particular user experience record.

QueryResponse
Response payload sent back to a physical web client. This response contains the record found based on the identiers present in a QueryRequest. The returned response will have a record, and sometimes details on normalization actions taken on the request that were necessary to make the request successful.

Record
Record is a single Chrome UX report data record. It contains use experience statistics for a single url pattern and set of dimensions.

RecordMethods
A builder providing access to all methods supported on record resources. It is not used directly, but through the ChromeUXReport hub.

RecordQueryRecordCall
Queries the Chrome User Experience for a single record for a given site. Returns a record that contains one or more metrics corresponding to performance data about the requested site.

UrlNormalization
Object representing the normalization actions taken to normalize a url to achieve a higher chance of successful lookup. These are simple automated changes that are taken when looking up the provided url_patten would be known to fail. Complex actions like following redirects are not handled.